perf(layout): 清理渲染警告
