+++ src/shared/components/layout/tags/Input/edit.tsx
@@ -19,7 +19,7 @@ export const TagInputEdit: TagComponent<TagInputProps> = React.memo((props) => {
   }, []);
 
   return (
-    <BaseTypeRow key={props.key}>
+    <BaseTypeRow>
       <LayoutCol span={6}>
         <TagLabel label={label} desc={props.desc} />
       </LayoutCol>
