+++ src/shared/components/layout/tags/TextArea/edit.tsx
@@ -24,7 +24,7 @@ export const TagTextAreaEdit: TagComponent<TagProps> = React.memo((props) => {
   );
 
   return (
-    <BaseTypeRow key={props.key}>
+    <BaseTypeRow>
       <LayoutCol span={6}>
         <TagLabel label={label} desc={props.desc} />
       </LayoutCol>
